Crate wasmtime_interface_types

A small crate to handle WebAssembly interface types in wasmtime.

Note that this is intended to follow the official proposal and is highly susceptible to change/breakage/etc.

Structs

ExportBinding
Representation of a binding of an exported function.

ModuleData
A data structure intended to hold a parsed representation of the wasm interface types of a module.

Enums

Value
The set of all possible WebAssembly Interface Types